Implement request factory.
